Extracting session controller methods
