--- app/controllers/concerns/with_referer.rb
@@ -1,5 +1,0 @@
-module WithReferer
-  def from_sessions?
-    params['controller'] == 'sessions'
-  end
-end
